Consolidated all instances of `box_for_tile` into one location.
